Remove circular dependency and hoist `MockData`
