+++ libs/mock-block-dock/src/datastore/mock-data.ts
@@ -1,0 +1,7 @@
+import { Entity, SubgraphTemporalAxes } from "@blockprotocol/graph";
+
+export type MockData = {
+  subgraphTemporalAxes: SubgraphTemporalAxes;
+  entities: Entity<true>[];
+  // linkedAggregationDefinitions: LinkedAggregationDefinition[];
+};
